refactor: Extract PodiumPlayers component from ResultPage
